+++ src/arize/experimental/integrations/whylabs_vanguard_ingestion/client.py
@@ -1,4 +1,5 @@
 # type: ignore[pb2]
+import datetime
 from typing import Dict, List, Optional, Union
 
 import pandas as pd
@@ -22,11 +23,11 @@ class IntegrationClient:
     def __init__(
         self,
         developer_key: str,
+        graphql_uri: str,
         api_key: Optional[str] = None,
         space_id: Optional[str] = None,
         space_key: Optional[str] = None,
         uri: Optional[str] = "https://api.arize.com/v1",
-        graphql_uri: Optional[str] = "https://app.arize.com/graphql",
         additional_headers: Optional[Dict[str, str]] = None,
         request_verify: Union[bool, str] = True,
         host: Optional[str] = None,
@@ -52,7 +53,7 @@ class IntegrationClient:
 
     def log_profile(
         self,
-        profile: DatasetProfile,
+        profile: DatasetProfile | DatasetProfileView,
         schema: BaseSchema,
         environment: Environments,
         model_id: str,
@@ -69,6 +70,9 @@ class IntegrationClient:
         num_rows: Optional[int] = None,
         kll_profile_view: Optional[DatasetProfileView] = None,
         n_kll_quantiles: Optional[int] = 200,
+        timestamp: Optional[
+            datetime.datetime
+        ] = None,  # Overrides timestamp value set in input schema
     ) -> requests.Response:
         """
         Logs a WhyLogs profile to the Arize API.
@@ -76,8 +80,8 @@ class IntegrationClient:
         The input must be a WhyLogs profile.
         """
         assert isinstance(
-            profile, DatasetProfile
-        ), f"Expected WhyLogs DatasetProfile, got {type(profile)}"
+            profile, (DatasetProfile, DatasetProfileView)
+        ), f"Expected WhyLogs DatasetProfile or DatasetProfileView, got {type(profile)}"
 
         if not self.model_exists(self._space_id, model_id, self._graphql_uri):
             raise ValueError(
@@ -90,6 +94,13 @@ class IntegrationClient:
             kll_profile_view=kll_profile_view,
             n_kll_quantiles=n_kll_quantiles,
         )
+
+        if timestamp:
+            synthetic_df["timestamp"] = timestamp
+            # Create a new schema with timestamp added
+            schema_dict = schema.asdict()
+            schema_dict["timestamp_column_name"] = "timestamp"
+            schema = Schema(**schema_dict)
 
         return self._client.log(
             dataframe=synthetic_df,
@@ -109,7 +120,7 @@ class IntegrationClient:
 
     def log_dataset_profile(
         self,
-        profile: DatasetProfile,
+        profile: DatasetProfile | DatasetProfileView,
         model_id: str,
         metrics_validation: Optional[List[Metrics]] = None,
         model_version: Optional[str] = None,
@@ -131,6 +142,9 @@ class IntegrationClient:
         tag_column_names: Optional[
             List[str]
         ] = None,  # List of columns to be used as tags
+        timestamp: Optional[
+            datetime.datetime
+        ] = None,  # Overrides timestamp value set in input schema
     ) -> requests.Response:
         """
         Logs a WhyLogs dataset-based profile to the Arize API.
@@ -138,8 +152,8 @@ class IntegrationClient:
         The input must be a WhyLogs profile.
         """
         assert isinstance(
-            profile, DatasetProfile
-        ), f"Expected WhyLogs DatasetProfile, got {type(profile)}"
+            profile, (DatasetProfile, DatasetProfileView)
+        ), f"Expected WhyLogs DatasetProfile or DatasetProfileView, got {type(profile)}"
 
         if not self.model_exists(self._space_id, model_id, self._graphql_uri):
             raise ValueError(
@@ -153,25 +167,25 @@ class IntegrationClient:
             n_kll_quantiles=n_kll_quantiles,
         )
 
-        schema = Schema(
-            feature_column_names=synthetic_df.columns.tolist(),
-            tag_column_names=tag_column_names,
-            prediction_label_column_name="ARIZE_PLACEHOLDER_STRING",
-            prediction_score_column_name="ARIZE_PLACEHOLDER_FLOAT",
-            actual_label_column_name="ARIZE_PLACEHOLDER_STRING",
-            actual_score_column_name="ARIZE_PLACEHOLDER_FLOAT",
-        )
+        schema_args = {
+            "feature_column_names": synthetic_df.columns.tolist(),
+            "tag_column_names": tag_column_names,
+            "prediction_label_column_name": prediction_label_column_name
+            or "ARIZE_PLACEHOLDER_STRING",
+            "prediction_score_column_name": prediction_score_column_name
+            or "ARIZE_PLACEHOLDER_FLOAT",
+            "actual_label_column_name": actual_label_column_name
+            or "ARIZE_PLACEHOLDER_STRING",
+            "actual_score_column_name": actual_score_column_name
+            or "ARIZE_PLACEHOLDER_FLOAT",
+        }
 
-        if timestamp_column_name:
-            schema.timestamp_column_name = timestamp_column_name
-        if prediction_score_column_name:
-            schema.prediction_score_column_name = prediction_score_column_name
-        if prediction_label_column_name:
-            schema.prediction_label_column_name = prediction_label_column_name
-        if actual_label_column_name:
-            schema.actual_label_column_name = actual_label_column_name
-        if actual_score_column_name:
-            schema.actual_score_column_name = actual_score_column_name
+        # Singular timestamp value for the synthetic dataset
+        if timestamp:
+            synthetic_df["timestamp"] = timestamp
+            schema_args["timestamp_column_name"] = "timestamp"
+
+        schema = Schema(**schema_args)
 
         synthetic_df["ARIZE_PLACEHOLDER_STRING"] = "ARIZE_PLACEHOLDER"
         synthetic_df["ARIZE_PLACEHOLDER_FLOAT"] = float("-inf")
@@ -251,9 +265,14 @@ class IntegrationClient:
             edge.get("model", {}).get("name") == model_id for edge in models
         )
 
-    def _extract_num_rows(self, profile: DatasetProfile) -> int:
+    def _extract_num_rows(
+        self, profile: DatasetProfile | DatasetProfileView
+    ) -> int:
         """Extracts number of rows from profile if not explicitly provided."""
-        profile_df = profile.view().to_pandas()
+        if isinstance(profile, DatasetProfileView):
+            profile_df = profile.to_pandas()
+        else:
+            profile_df = profile.view().to_pandas()
 
         if "counts/n" not in profile_df.columns:
             raise ValueError("Profile is missing required 'counts/n' column")
@@ -270,12 +289,16 @@ class IntegrationClient:
 
     def _generate_synthetic_dataset(
         self,
-        profile: DatasetProfile,
+        profile: DatasetProfile | DatasetProfileView,
         num_rows: int,
         kll_profile_view: Optional[DatasetProfileView] = None,
         n_kll_quantiles: Optional[int] = 200,
     ) -> pd.DataFrame:
-        profile_df = profile.view().to_pandas()
+        if isinstance(profile, DatasetProfileView):
+            profile_df = profile.to_pandas()
+        else:
+            profile_df = profile.view().to_pandas()
+
         return self._profile_adapter.generate(
             profile_df,
             num_rows=num_rows,