+++ src/arize/experimental/online_tasks/__init__.py
@@ -1,0 +1,3 @@
+from .dataframe_preprocessor import extract_nested_data_to_column
+
+__all__ = ["extract_nested_data_to_column"]